End of branch changes
Rename the Number class to BaseClass for Number Template Class
Refactoring code and cleaned the warnings.
Added Number class files for template class.
Moved calculating operations operator bodies to operation's .cpp file
