+++ src/cpp-files/Number/Decimal/multiplying.cpp
@@ -1,0 +1,80 @@
+//
+// Created by xaprier on 2/13/23.
+//
+#include "header-files/Number/Decimal.h"
+
+Decimal Decimal::operator*(const Decimal& second) {
+	// creating return object/converting other object
+	Decimal result(*this);
+
+	// calculate the multiplying of decimal values
+	result.num = this->mul(result.num, second.num);
+
+	// remove the last indexes if it is 0
+	while (result.num.at(result.num.length() - 1) == '0' && result.num.at(result.num.length() - 2) != '.')
+		result.num = result.num.substr(0, result.num.length() - 1);
+
+	return result;
+}
+
+Decimal Decimal::operator*(const Binary& b) {
+	// creating return object/converting other object
+	Decimal result(*this), second(b);
+
+	// calculate the multiplying of decimal values
+	result.num = this->mul(result.num, second.num);
+
+	// remove the last indexes if it is 0
+	while (result.num.at(result.num.length() - 1) == '0' && result.num.at(result.num.length() - 2) != '.')
+		result.num = result.num.substr(0, result.num.length() - 1);
+
+	return result;
+}
+
+Decimal Decimal::operator*(const Octal& o) {
+	// creating return object/converting other object
+	Decimal result(*this), second(o);
+
+	// calculate the multiplying of decimal values
+	result.num = this->mul(result.num, second.num);
+
+	// remove the last indexes if it is 0
+	while (result.num.at(result.num.length() - 1) == '0' && result.num.at(result.num.length() - 2) != '.')
+		result.num = result.num.substr(0, result.num.length() - 1);
+
+	return result;
+}
+
+Decimal Decimal::operator*(const Hexadecimal& h) {
+	// creating return object/converting other object
+	Decimal result(*this), second(h);
+
+	// calculate the multiplying of decimal values
+	result.num = this->mul(result.num, second.num);
+
+	// remove the last indexes if it is 0
+	while (result.num.at(result.num.length() - 1) == '0' && result.num.at(result.num.length() - 2) != '.')
+		result.num = result.num.substr(0, result.num.length() - 1);
+
+	return result;
+}
+
+Decimal &Decimal::operator*=(const Decimal &d) {
+	*this = *this * d;
+	return *this;
+}
+
+Decimal &Decimal::operator*=(const Binary &b) {
+	*this = *this * b;
+	return *this;
+}
+
+Decimal &Decimal::operator*=(const Octal &o) {
+	*this = *this * o;
+	return *this;
+}
+
+Decimal &Decimal::operator*=(const Hexadecimal &h) {
+	*this = *this * h;
+	return *this;
+}
