+++ src/header-files/Number/Octal.h
@@ -1,12 +1,12 @@
 #ifndef OCTAL_H
 #define OCTAL_H
 
-#include "Number.h"
+#include "BaseClass.h"
 #include "Decimal.h"
 #include "Binary.h"
 #include "Hexadecimal.h"
 
-class Octal : public Number {
+class Octal : public BaseClass {
   public:
 	explicit Octal(std::string num = "0.0");
 	explicit Octal(const Hexadecimal &h);
@@ -19,35 +19,35 @@ class Octal : public Number {
 	Binary toBin() override;
 	Hexadecimal toHex() override;
 
-	operator std::string() override;
+	explicit operator std::string() override;
 
 	friend std::ostream &operator<<(std::ostream &output, const Octal &o);
 
 	// operator overloadings of summing
-	Octal operator+(Octal o);
-	Octal operator+(Binary b);
-	Octal operator+(Decimal second);
-	Octal operator+(Hexadecimal h);
+	Octal operator+(const Octal& o);
+	Octal operator+(const Binary& b);
+	Octal operator+(const Decimal& second);
+	Octal operator+(const Hexadecimal& h);
 	// operator overloadings of substraction
-	Octal operator-(Octal o);
-	Octal operator-(Binary b);
-	Octal operator-(Decimal second);
-	Octal operator-(Hexadecimal h);
+	Octal operator-(const Octal& o);
+	Octal operator-(const Binary& b);
+	Octal operator-(const Decimal& second);
+	Octal operator-(const Hexadecimal& h);
 	// operator overloadings of multiply
-	Octal operator*(Octal o);
-	Octal operator*(Binary b);
-	Octal operator*(Decimal second);
-	Octal operator*(Hexadecimal h);
+	Octal operator*(const Octal& o);
+	Octal operator*(const Binary& b);
+	Octal operator*(const Decimal& second);
+	Octal operator*(const Hexadecimal& h);
 	// operator overloadings of divide
-	Octal operator/(Octal o);
-	Octal operator/(Binary b);
-	Octal operator/(Decimal second);
-	Octal operator/(Hexadecimal h);
+	Octal operator/(const Octal& o);
+	Octal operator/(const Binary& b);
+	Octal operator/(const Decimal& second);
+	Octal operator/(const Hexadecimal& h);
 	// operator overloadings of remainder
-	Octal operator%(Octal o);
-	Octal operator%(Binary b);
-	Octal operator%(Decimal second);
-	Octal operator%(Hexadecimal h);
+	Octal operator%(const Octal& o);
+	Octal operator%(const Binary& b);
+	Octal operator%(const Decimal& second);
+	Octal operator%(const Hexadecimal& h);
 	// assignment operator overloadings
 	Octal &operator=(const Octal &o);
